Move ExternalDataInstance.java to the correct location (where it is in commcare-core).
